Move getLumosOptions function to Options.ts file.
